fix and refactor dependency injection for userService and PrismaClient

+++ invoice-backend/src/services/user.service.ts
@@ -1,27 +1,27 @@
 import { inject, injectable } from "inversify";
-import { PrismaUserRepo } from "../repositories/implementations/prismaUserRepo";
 import {
   UserEntity,
   CreateUserDTO,
   UserDTO,
-  LoginResponseDTO,
-  CreateUserArgs,
   UserIdAndRole,
 } from "../constants/types";
 import bcrypt from "bcrypt";
 import { IUserRepo } from "../repositories/userRepo";
 import { validateUserCreate, validateUserList } from "../utils/utils";
-import { SECRET } from "../config/server.config";
-import jwt from "jsonwebtoken";
 import {
   InternalServerException,
   NotFoundException,
-  UnauthorizedException,
+  ValidationException,
 } from "../config/exception.config";
+import TYPES from "@/constants/identifiers";
 
 @injectable()
 export class UserService {
-  constructor(@inject(PrismaUserRepo) private readonly userRepo: IUserRepo) {}
+  constructor(
+    @inject(TYPES.IUserRepo) private readonly userRepo: IUserRepo,
+    @inject(TYPES.UserContext)
+    private readonly userContext: UserIdAndRole,
+  ) {}
 
   createUser = async (args: CreateUserDTO): Promise<UserDTO> => {
     const validatedArgs = validateUserCreate(args);
@@ -58,6 +58,15 @@ export class UserService {
   };
 
   getUsers = async () => {
+    if (!this.userContext) {
+      throw new ValidationException("Unauthorized");
+    }
+
+    const { role, id } = this.userContext;
+
+    if (!id) {
+      throw new ValidationException("Unauthorized");
+    }
     try {
       const userList = await this.userRepo.getAllUsers();
       const validatedUserList: UserDTO[] = validateUserList(userList);
@@ -69,6 +78,15 @@ export class UserService {
   };
 
   getUser = async (id: string) => {
+    if (!this.userContext) {
+      throw new ValidationException("Unauthorized");
+    }
+
+    const { role } = this.userContext;
+
+    if (!id) {
+      throw new ValidationException("Unauthorized");
+    }
     const user = await this.userRepo.getUserById(id);
 
     if (!user) {
@@ -83,48 +101,38 @@ export class UserService {
     return userDTO;
   };
 
-  getUserSafely = async (id: string) => {
-    const user = await this.userRepo.getUserSafely(id);
+  getUserByIdSafely = async (id: string) => {
+    const user = await this.userRepo.getUserByIdSafely(id);
     return user;
   };
 
-  // login = async (
-  //   username: string,
-  //   password: string,
-  // ): Promise<LoginResponseDTO> => {
-  //   const user = await this.userRepo.getUserSafely(username);
-
-  //   if (!user) {
-  //     throw new NotFoundException("User not found");
-  //   }
-
-  //   const match = await bcrypt.compare(password, user.passwordHash);
-
-  //   if (!match) {
-  //     throw new UnauthorizedException("Invalid username or password");
-  //   }
-
-  //   const tokenPayload = {
-  //     id: user.id,
-  //     username: user.username,
-  //   };
-
-  //   const userDTO: UserDTO = {
-  //     id: user.id,
-  //     name: user.name ?? "",
-  //     username: user.username,
-  //   };
-
-  //   const token = jwt.sign(tokenPayload, SECRET, { expiresIn: "1h" });
-  //   const loginResponse: LoginResponseDTO = { token: token, user: userDTO };
-  //   return loginResponse;
-  // };
-
   deleteUsers = async (): Promise<boolean> => {
+    if (!this.userContext) {
+      throw new ValidationException("Unauthorized");
+    }
+
+    const { role, id } = this.userContext;
+
+    if (!id) {
+      throw new ValidationException("Unauthorized");
+    }
     return await this.userRepo.deleteAllUsers();
   };
 
   deleteUsersKeepAdmin = async (): Promise<boolean> => {
+    if (!this.userContext) {
+      throw new ValidationException("Unauthorized");
+    }
+
+    const { role, id } = this.userContext;
+
+    if (!id) {
+      throw new ValidationException("Unauthorized");
+    }
+
+    if (role !== "ADMIN") {
+      throw new ValidationException("Unauthorized - must be ADMIN");
+    }
     return await this.userRepo.deleteAllUsersKeepAdmin();
   };
 }
